Make split query dataset specific
