+++ snuba/datasets/events/__init__.py
@@ -4,6 +4,7 @@ from datetime import timedelta
 from typing import Mapping, Sequence
 
 from snuba import state
+from snuba.datasets.events import split
 from snuba.clickhouse.columns import (
     Array,
     ColumnSet,
@@ -18,7 +19,7 @@ from snuba.clickhouse.columns import (
 from snuba.datasets import TimeSeriesDataset
 from snuba.datasets.dataset_schemas import DatasetSchemas
 from snuba.datasets.table_storage import TableWriter, KafkaStreamLoader
-from snuba.datasets.events_processor import EventsProcessor
+from snuba.datasets.events.events_processor import EventsProcessor
 from snuba.datasets.schemas.tables import ReplacingMergeTreeSchema
 from snuba.query.extensions import (
     PerformanceExtension,
@@ -405,3 +406,6 @@ class EventsDataset(TimeSeriesDataset):
 
     def get_prewhere_keys(self) -> Sequence[str]:
         return ['event_id', 'issue', 'tags[sentry:release]', 'message', 'environment', 'project_id']
+
+    def split_query(self, query_func, *args, **kwargs):
+        return split.split_query(query_func)(self, *args, **kwargs)
